fix bulk operations operating on closed file in ImageSetAPI
When trying to perfom a bulk operation using either a list or file
name you get the following ValueError:

    ValueError: I/O operation on closed file.

The context manager `_normalize_items_file` creates a temp file (or
read an existing file) and yield the open file.
The request to the API was not executed inside the context so the file
is closed before the request is actually made.

Fix was to move the call into the context.